+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- fix duplicated extensions in build artifacts file names ([#12](https://github.com/seaofvoices/generator-luau/pull/12))
 - fix `build-roblox-asset.sh` script ([#11](https://github.com/seaofvoices/generator-luau/pull/11))
 - fix `remove-tests.sh` script to support both `lua` and `luau` extensions ([#10](https://github.com/seaofvoices/generator-luau/pull/10))
 - move Roblox `test-place.rbxl` file into the `temp` directory ([#9](https://github.com/seaofvoices/generator-luau/pull/9))